+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/lightspeed/trex/stage_manager/plugin/interaction/usd/all_categories.py
@@ -52,7 +52,6 @@ class AllCategoriesInteractionPlugin(_StageManagerUSDInteractionPlugin):
             "CustomTagsWidgetPlugin",
             "DeleteRestoreActionWidgetPlugin",
             "FocusInViewportActionWidgetPlugin",
-            "GeometryPrimsFilterPlugin",
             "IsCaptureStateWidgetPlugin",
             "IsCategoryHiddenStateWidgetPlugin",
             "IsVisibleActionWidgetPlugin",