+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_lights.py
@@ -59,6 +59,7 @@ class AllLightsInteractionPlugin(_StageManagerUSDInteractionPlugin):
         default=[
             "AssignCategoryActionWidgetPlugin",
             "CustomTagsWidgetPlugin",
+            "DeleteRestoreActionWidgetPlugin",
             "FocusInViewportActionWidgetPlugin",
             "IsCaptureStateWidgetPlugin",
             "IsCategoryHiddenStateWidgetPlugin",